+++ package.json
@@ -47,20 +47,20 @@
     "proxy-action:proxy-action-free-collect": "env-cmd -f .env ts-node ./src/proxy-action/proxy-action-free-collect",
     "proxy-action:proxy-action-free-follow": "env-cmd -f .env ts-node ./src/proxy-action/proxy-action-free-follow",
     "publications:comment": "env-cmd -f .env ts-node ./src/publications/comment",
-    "publications:comment-encrypted": "env-cmd -f .env ts-node ./src/publications/comment-encrypted",
-    "publications:comment-encrypted-gasless": "env-cmd -f .env ts-node ./src/publications/comment-encrypted-gasless",
+    "publications:comment-gated": "env-cmd -f .env ts-node ./src/publications/comment-gated",
+    "publications:comment-gated-gasless": "env-cmd -f .env ts-node ./src/publications/comment-gated-gasless",
     "publications:comment-gasless": "env-cmd -f .env ts-node ./src/publications/comment-gasless",
     "publications:mirror": "env-cmd -f .env ts-node ./src/publications/mirror",
     "publications:mirror-gasless": "env-cmd -f .env ts-node ./src/publications/mirror-gasless",
     "publications:get-publication": "env-cmd -f .env ts-node ./src/publications/get-publication",
-    "publications:get-publication-encrypted": "env-cmd -f .env ts-node ./src/publications/get-publication-encrypted",
+    "publications:get-publication-gated": "env-cmd -f .env ts-node ./src/publications/get-publication-gated",
     "publications:get-publications": "env-cmd -f .env ts-node ./src/publications/get-publications",
     "publications:hide-publication": "env-cmd -f .env ts-node ./src/publications/hide-publication",
     "publications:post-gasless": "env-cmd -f .env ts-node ./src/publications/post-gasless",
-    "publications:post-encrypted-gasless": "env-cmd -f .env ts-node ./src/publications/post-encrypted-gasless",
+    "publications:post-gated-gasless": "env-cmd -f .env ts-node ./src/publications/post-gated-gasless",
     "publications:post": "env-cmd -f .env ts-node ./src/publications/post",
     "publications:post-attach": "env-cmd -f .env ts-node ./src/publications/post-attach",
-    "publications:post-encrypted": "env-cmd -f .env ts-node ./src/publications/post-encrypted",
+    "publications:post-gated": "env-cmd -f .env ts-node ./src/publications/post-gated",
     "publications:profile-publications-for-sale": "env-cmd -f .env ts-node ./src/publications/profile-publications-for-sale",
     "publications:who-collected-publication": "env-cmd -f .env ts-node ./src/publications/who-collected-publication",
     "reaction:add": "env-cmd -f .env ts-node ./src/reaction/add-reaction",
